fix pkg names
